tweak: change "distance" value from integer to floating point

+++ src/main/java/com/xyz/gps/poi/domain/dto/POIDto.java
@@ -33,7 +33,7 @@ public class POIDto {
     @PositiveOrZero(message = "Coordinate 'y' must be a positive value.")
     private Long y;
 
-    private Long distance;
+    private Double distance;
 
     public POIDto(POI poi) {
         name = poi.getName();
